requirements.txt
anaconda==0.0.1.1
asgiref==3.7.2
decorator==5.1.1
django-js-asset==2.1.0
djangorestframework==3.14.0
Pillow==10.0.0
pytz==2023.3.post1
self==2020.12.3
sqlparse==0.4.4
tzdata==2023.3
